Move InputNumber helpers + support non gouped values
